Move edit prediction layout functions to Editor
